Upgrade to Laravel 5.2 / PhantomJS v2

+++ app/Http/Controllers/ApiKeysController.php
@@ -11,28 +11,6 @@ use Screeenly\Http\Requests\ApiKeyRequest;
 class ApiKeysController extends Controller
 {
     /**
-     * Display a listing of the resource.
-     *
-     * @return Response
-     */
-    public function index()
-    {
-        $apikeys = auth()->user()->apikeys()->latest()->get();
-
-        return view('app.apikeys.index', compact('apikeys'));
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return Response
-     */
-    public function create()
-    {
-        return view('app.apikeys.create');
-    }
-
-    /**
      * Store a newly created resource in storage.
      *
      * @return Response
@@ -68,6 +46,10 @@ class ApiKeysController extends Controller
      */
     public function edit(ApiKey $apikey)
     {
+        if ($apikey->user->id != auth()->id()) {
+            return redirect("/");
+        }
+
         return view('app.apikeys.edit', compact('apikey'));
     }
 
@@ -79,6 +61,10 @@ class ApiKeysController extends Controller
      */
     public function update(ApiKeyRequest $request, ApiKey $apikey)
     {
+        if ($apikey->user->id != auth()->id()) {
+            return redirect("/");
+        }
+
         $apikey->update($request->all());
 
         return redirect()->route('app.dashboard')->withMessage("Key {$apikey->name} saved!");
@@ -92,6 +78,10 @@ class ApiKeysController extends Controller
      */
     public function destroy(ApiKey $apikey)
     {
+        if ($apikey->user->id != auth()->id()) {
+            return redirect("/");
+        }
+
         $apikey->delete();
 
         return redirect()->back()->withMessage("API Key deleted");
